Extracted VariantMongoWriter to it's own configuration class.
