+++ src/main/java/uk/ac/ebi/eva/pipeline/Application.java
@@ -34,6 +34,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class Application {
 
     private static final Logger log = LoggerFactory.getLogger(Application.class);
+    public static final String VARIANT_WRITER_MONGO_PROFILE = "variant-writer-mongo";
     public static final String VARIANT_ANNOTATION_MONGO_PROFILE = "variant-annotation-mongo";
 
     public static void main(String[] args) throws Exception {
