+++ src/main/resources/application-mongo.properties
@@ -1,1 +1,1 @@
-spring.profiles.include=variant-annotation-mongo
+spring.profiles.include=variant-writer-mongo, variant-annotation-mongo
