+++ src/test/resources/genotyped-vcf.properties
@@ -1,3 +1,4 @@
+spring.profiles.active=variant-writer-mongo, variant-annotation-mongo
 input.vcf=/small20.vcf.gz
 input.vcf.id=1
 input.vcf.aggregation=NONE
