+++ src/test/resources/variant-aggregated.properties
@@ -1,4 +1,4 @@
-spring.profiles.active=variant-annotation-mongo
+spring.profiles.active=variant-writer-mongo, variant-annotation-mongo
 input.vcf=/aggregated.vcf.gz
 input.vcf.id=1
 input.vcf.aggregation=BASIC
